Refactoring APv1 and APv2 to utilize the same memory AP bus-specific configurations
