+++ probe-rs/src/architecture/arm/ap/memory_ap/amba_ahb3.rs
@@ -1,6 +1,5 @@
 use crate::architecture::arm::{
-    ap::{define_ap_register, ApRegister, CFG},
-    ap_v1::{AccessPortType, ApAccess, ApRegAccess},
+    ap::{define_ap_register, AccessPortType, ApAccess, ApRegAccess, ApRegister, CFG},
     ArmError, DapAccess, FullyQualifiedApAddress, RegisterParseError,
 };
 
@@ -11,14 +10,14 @@ use super::{AddressIncrement, DataSize};
 /// The memory AP can be used to access a memory-mapped
 /// set of debug resources of the attached system.
 #[derive(Debug)]
-pub struct AmbaAhb5 {
+pub struct AmbaAhb3 {
     address: FullyQualifiedApAddress,
     csw: CSW,
     cfg: CFG,
 }
 
-impl AmbaAhb5 {
-    /// Creates a new AmbaAhb5 with `address` as base address.
+impl AmbaAhb3 {
+    /// Creates a new AmbaAhb3 with `address` as base address.
     pub fn new<P: DapAccess>(
         probe: &mut P,
         address: FullyQualifiedApAddress,
@@ -32,6 +31,7 @@ impl AmbaAhb5 {
             DbgSwEnable: true,
             HNONSEC: !csw.SPIDEN,
             MasterType: true,
+            Cacheable: true,
             Privileged: true,
             Data: true,
             AddrInc: AddressIncrement::Single,
@@ -42,7 +42,7 @@ impl AmbaAhb5 {
     }
 }
 
-impl super::MemoryApType for AmbaAhb5 {
+impl super::MemoryApType for AmbaAhb3 {
     type CSW = CSW;
 
     fn status<P: ApAccess + ?Sized>(&mut self, probe: &mut P) -> Result<CSW, ArmError> {
@@ -84,20 +84,20 @@ impl super::MemoryApType for AmbaAhb5 {
     }
 
     fn supports_only_32bit_data_size(&self) -> bool {
-        // Amba AHB5 must support word, half-word and byte size transfers.
+        // Amba AHB3 must support word, half-word and byte size transfers.
         false
     }
 }
 
-impl AccessPortType for AmbaAhb5 {
+impl AccessPortType for AmbaAhb3 {
     fn ap_address(&self) -> &FullyQualifiedApAddress {
         &self.address
     }
 }
 
-impl ApRegAccess<CSW> for AmbaAhb5 {}
+impl ApRegAccess<CSW> for AmbaAhb3 {}
 
-super::attached_regs_to_mem_ap!(memory_ap_regs => AmbaAhb5);
+super::attached_regs_to_mem_ap!(memory_ap_regs => AmbaAhb3);
 
 define_ap_register!(
     /// Control and Status Word register
@@ -119,11 +119,13 @@ define_ap_register!(
         ///
         /// Support of this function is implementation defined.
         MasterType: bool,           // [29]
-        /// `HPROT[6:3]`.
+        /// Drives `HPROT[4]`, Allocate.
         ///
-        /// - bit 3 controls bits `HPROT`’s bit 3, 4 and 6
-        /// - `HPROT[5]` is tied to 0
-        CombinedHPROT346: bool,     // [27]
+        /// `HPROT[4]` is an Armv5 extension to AHB. For more information, see the Arm1136JF-S™ and
+        /// Arm1136J-S ™ Technical Reference Manual.
+        Allocate: bool,             // [28]
+        /// `HPROT[3]`
+        Cacheable: bool,            // [27]
         /// `HPROT[2]`
         Bufferable: bool,           // [26]
         /// `HPROT[1]`
@@ -164,7 +166,8 @@ define_ap_register!(
         DbgSwEnable: ((value >> 31) & 0x01) != 0,
         HNONSEC:    ((value >> 30) & 0x01) != 0,
         MasterType: ((value >> 29) & 0x01) != 0,
-        CombinedHPROT346: ((value >> 27) & 0x01) != 0,
+        Allocate:   ((value >> 28) & 0x01) != 0,
+        Cacheable:  ((value >> 27) & 0x01) != 0,
         Bufferable: ((value >> 26) & 0x01) != 0,
         Privileged: ((value >> 25) & 0x01) != 0,
         Data:       ((value >> 24) & 0x01) != 0,
@@ -173,12 +176,13 @@ define_ap_register!(
         DeviceEn:   ((value >> 6) & 0x01) != 0,
         AddrInc: AddressIncrement::from_u8(((value >> 4) & 0x03) as u8).ok_or_else(|| RegisterParseError::new("CSW", value))?,
         Size: DataSize::try_from((value & 0x07) as u8).map_err(|_| RegisterParseError::new("CSW", value))?,
-        _reserved_bits: value & 0x107F_FF08,
+        _reserved_bits: value & 0x007F_FF08,
     }),
     to: value => (u32::from(value.DbgSwEnable) << 31)
     | (u32::from(value.HNONSEC      ) << 30)
     | (u32::from(value.MasterType   ) << 29)
-    | (u32::from(value.CombinedHPROT346) << 27)
+    | (u32::from(value.Allocate     ) << 28)
+    | (u32::from(value.Cacheable    ) << 27)
     | (u32::from(value.Bufferable   ) << 26)
     | (u32::from(value.Privileged   ) << 25)
     | (u32::from(value.Data         ) << 24)
