+++ probe-rs/src/architecture/arm/mod.rs
@@ -2,8 +2,6 @@
 
 #[macro_use]
 pub mod ap;
-pub mod ap_v1;
-pub mod ap_v2;
 pub(crate) mod assembly;
 pub(crate) mod communication_interface;
 pub mod component;
@@ -17,7 +15,7 @@ mod traits;
 
 pub use self::core::{armv6m, armv7a, armv7m, armv8a, armv8m, Dump};
 use self::{
-    ap_v1::AccessPortError,
+    ap::AccessPortError,
     dp::DebugPortError,
     memory::romtable::RomTableError,
     sequences::ArmDebugSequenceError,
