+++ probe-rs/src/architecture/arm/sequences.rs
@@ -21,7 +21,7 @@ use crate::{
 };
 
 use super::{
-    ap_v1::AccessPortError,
+    ap::AccessPortError,
     armv6m::Demcr,
     communication_interface::{DapProbe, Initialized},
     component::{TraceFunnel, TraceSink},
